Removes editor lib
